% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_DF.R
\name{Average_DF}
\alias{Average_DF}
\title{Average multiple data frames in a list}
\usage{
Average_DF(DF_list, unique_comb)
}
\arguments{
\item{DF_list}{a list of data frames to be averaged}

\item{unique_comb}{a list of column names that are used as the unique identifiers}
}
\value{
A dataframe which contains average data for each unique combination
}
\description{
Takes any number of data frames and averages them
}
