% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct_offset.R
\name{Correct_offset}
\alias{Correct_offset}
\title{Correct DC offset in unrectified electromyographical data}
\usage{
Correct_offset(x)
}
\arguments{
\item{x}{a vector of electromyographical voltage data}
}
\value{
A vector with corrected data, where the mean is now zero
}
\description{
Takes unrectified electromyographical data, and removes DC offset so data is centred around zero voltage
}
