% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Critically_damped.R
\name{Critically_damped}
\alias{Critically_damped}
\title{Critically damped filter}
\usage{
Critically_damped(raw_data, sampling_frequency = 1000,
  cutoff_frequency = 20, filter_passes = 10)
}
\arguments{
\item{raw_data}{vector of raw numeric data to be filtered}

\item{sampling_frequency}{numeric value indicating the sampling frequency fo the data to be filtered. Defaults to 1000.}

\item{cutoff_frequency}{numeric value indicating the cutoff frequency of the low-pass filter. Defaults to 20Hz.}

\item{filter_passes}{numeric value indicating the number of passes to be made by the filter. Defaults to 10.}
}
\description{
This function allows you to low-pass filter data via a critically damped filter
}
