% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_envelope.R
\name{Linear_envelope}
\alias{Linear_envelope}
\title{Process raw electromyographical data to produce a linear envelope}
\usage{
Linear_envelope(x, high = 20, low = 6, final_order = 4,
  sample_frequency = 1000)
}
\arguments{
\item{x}{a vector of electromyographical voltage data}

\item{high}{numeric value indicating the cutoff frequency of the high-pass filter. Defaults to 20Hz.}

\item{low}{numeric value indicating the cutoff frequency of the low-pass filter. Defaults to 6Hz.}

\item{final_order}{numeric value indicating the final order of the filter. Defaults to 4.}

\item{sample_frequency}{numeric value indicating the sampling freuncy fo the data to be filtered. Defaults to 1000}
}
\value{
vector containing linear envelope of the electromyographical data
}
\description{
Linear envelope of electromyographical data via high-pass filtering, full wave rectify and low-pass filtering the raw data
}
