% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalise_EMG.R
\name{Normalise_EMG}
\alias{Normalise_EMG}
\title{Normalise EMG to reference value.}
\usage{
Normalise_EMG(x, maxes)
}
\arguments{
\item{x}{a dataframe where each column provide a different channel of electromyographical voltage data.}

\item{maxes}{a dataframe or list of dataframes containing electromyographical data of which you want to find the maximum reference value. Dataframes must have same columns as x.}
}
\value{
dataframe with normalised electromyographical voltage data.
}
\description{
Normalise EMG to reference value.
}
