% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMG_read.R
\name{read.EMG}
\alias{read.EMG}
\title{Read a csv file output from VICON Nexus and extracts the electromyographical data}
\usage{
read.EMG(x, from.col = 21, to.col = 28, muscles = c("EMG_1", "EMG_2",
  "EMG_3", "EMG_4", "EMG_5", "EMG_6", "EMG_7", "EMG_8"), skip_rows = 4)
}
\arguments{
\item{x}{csv file output from VICON nexus that contains electromyographical data}

\item{from.col}{numeric value indicating the first column at which electromyographical data appears}

\item{to.col}{numeric value indicating the last column at which electromyographical data appears}

\item{muscles}{a vector of strings used to name each channel. Must be the same length as difference between from.col an to.col}

\item{skips_rows}{numeric value indicating the number of rows to skip until the first row of EMG data. Defaults to 4.}
}
\value{
Dataframe containing the electromyographical data extracted from the csv
}
\description{
Read a csv file output from VICON Nexus and extracts the electromyographical data
}
