% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residual_analysis.R
\name{residual_analysis}
\alias{residual_analysis}
\title{Process raw data to determine residuals between the raw and filtered signal using a range of low pass cutoff frequencies for butterworth filter}
\usage{
residual_analysis(raw_signal, cutoff_range = c(0, 20), sample_freq = 1000,
  final_order = 4, interval = 0.01)
}
\arguments{
\item{raw_signal}{vector of raw values of which you wish to determine the optimal cutoff frequency.}

\item{cutoff_range}{numeric vector of format c(lower,upper) containing the lower and upper cutoff frequency limit. Defaults to c(0,20).}

\item{sample_freq}{numeric value indicating the frequency of the signal. Defaults to 1000Hz.}

\item{final_order}{numeric value indicating the final order of the filter. Defaults to 4.}

\item{interval}{numeric value indicating the increment by which the cutoff frequency will increase within the limits specified by cutoff_range. Defaults to 0.01.}
}
\value{
dataframe containing the cutoff frequencies and the respective residual at the given cutoff frequency.
}
\description{
Process raw data to determine residuals between the raw and filtered signal using a range of low pass cutoff frequencies for butterworth filter.
}
