% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residual_cutoff.R
\name{residual_cut}
\alias{residual_cut}
\title{Process cutoff frequency and residual data to determine optimal low pass cutoff frequency for butterworth filter}
\usage{
residual_cut(residual_analysis, Hz_range, interval = 0.01)
}
\arguments{
\item{residual_analysis}{dataframe where first column contains cutoff frequencies, and the second column contains the respective residual between the raw and filtered signal. Can be generated using the Residual_analysis function.}

\item{Hz_range}{numeric vector of format c(lower,upper) containing the lower and upper cutoff frequency limit which identifies the linear region of the curve.}

\item{interval}{numeric value indicating the increment by which the cutoff frequency will increase within the limits specified by cutoff_range. Defaults to 0.01.}
}
\value{
numeric value of the recommended cutoff frequency.
}
\description{
Process residual_analysis frequency and residual data to determine optimal low pass cutoff frequency for butterworth filter.
}
