% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMS_error.R
\name{RMS_error}
\alias{RMS_error}
\title{Compute the root-mean-square difference between two vectors of the same length}
\usage{
RMS_error(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector of the same length as x}
}
\value{
numeric value of the root-mean-square difference between vectors x and y
}
\description{
Compute the root-mean-square difference between two vectors of the same length
}
