% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_EMG_mot.R
\name{write_EMG_mot}
\alias{write_EMG_mot}
\title{Compute the root-mean-square difference between two vectors of the same length}
\usage{
write_EMG_mot(x, name = "Normalised EMG Envelopes", filename = "EMG")
}
\arguments{
\item{x}{data frame containing the data to be written to a .mot file}

\item{name}{string for the name parameter printed in the .mot file. See example .MOT files provided with the OpenSim package.}

\item{filename}{string containing the desired filename for the .mot file. Note the ".mot" is note required in this string}
}
\value{
writes the data frame to a .mot file
}
\description{
Compute the root-mean-square difference between two vectors of the same length
}
